[commit info]commit for delay time show and delay test

+++ server_pc/src/cn/hisdar/cr/communication/data/AbstractData.java
@@ -37,7 +37,7 @@ public abstract class AbstractData {
 		// the length of long is 8 bytes
 		long number = 0;
 		for (int i = 0; i < bytesData.length; i++) {
-			number |= ((bytesData[i] & 0xFF) << (i * 8));
+			number |= (((long)(bytesData[i] & 0xFF)) << (i * 8));
 		}
 
 		return number;
